+++ pkg/bot/fuel_handler.go
@@ -1,6 +1,7 @@
 package bot
 
 import (
+	"context"
 	"fmt"
 	"strings"
 	"time"
@@ -212,11 +213,23 @@ func (b *fuelBot) allStructuresMessage(structures []structureData) *discordgo.Me
 		fields    []*discordgo.MessageEmbedField
 		fuelTotal float64
 	)
+
 	for _, structureData := range structures {
 		structureType := structureByTypeID(structureData.CorporationData.TypeId)
+		// Add symbols for time ranges for fuel remaining. Green = OK
+		symbol := ":green_square:"
+		// < 7 days = orange
+		if time.Until(structureData.CorporationData.FuelExpires) < 7*time.Hour*24 {
+			symbol = ":orange_square:"
+		}
+		// < 1 day = red
+		if time.Until(structureData.CorporationData.FuelExpires) < 1*time.Hour*24 {
+			symbol = ":red_square:"
+		}
 
 		field := &discordgo.MessageEmbedField{
-			Name: fmt.Sprintf("%s (%s)",
+			Name: fmt.Sprintf("%s %s (%s)",
+				symbol,
 				structureData.UniverseData.Name,
 				structureType.Name,
 			),
@@ -237,11 +250,28 @@ func (b *fuelBot) allStructuresMessage(structures []structureData) *discordgo.Me
 		}
 		fields = append(fields, field)
 	}
+
+	month := 30.0
+	dailyFuelMsg := fmt.Sprintf("**Daily**: %.0f", fuelTotal)
+	monthlyFuelMsg := fmt.Sprintf("**Monthly**: %.0f", fuelTotal*month)
+
+	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
+	defer cancel()
+	fuelPrices, err := b.estFuelPrice(ctx)
+	if err != nil {
+		b.log.Errorw("unable to estimate fuel prices", "error", err)
+	}
+	fuelDailyPrices := formatFuelPrices(fuelTotal, fuelPrices)
+	fuelMonthlyPrices := formatFuelPrices(fuelTotal*month, fuelPrices)
+
+	finishedDailyMsg := fmt.Sprintf("%s %s", dailyFuelMsg, fuelDailyPrices)
+	finishedMonthlyMsg := fmt.Sprintf("%s %s", monthlyFuelMsg, fuelMonthlyPrices)
+
 	fields = append(fields, &discordgo.MessageEmbedField{
-		Name: "--- Total fuel ---",
-		Value: fmt.Sprintf("**Daily**: %.0f \n**Monthly**: %.0f",
-			fuelTotal,
-			fuelTotal*30,
+		Name: ":ice_cube: Total fuel",
+		Value: fmt.Sprintf("%s \n%s",
+			finishedDailyMsg,
+			finishedMonthlyMsg,
 		),
 	})
 
@@ -281,6 +311,94 @@ func (b *fuelBot) structureFuelPerDay(structure structureData, structureType str
 		}
 	}
 	return acc * 24
+}
+
+func (b *fuelBot) estFuelPrice(ctx context.Context) (map[int32]float64, error) {
+	var (
+		regionID int32 = 10000002 // The Forge (Jita)
+		typeIDs        = []int32{
+			4247, // "Helium Fuel Block"
+			4246, // "Hydrogen Fuel Block"
+			4051, // "Nitrogen Fuel Block"
+			4312, // "Oxygen Fuel Block"
+		}
+	)
+
+	var out = make(map[int32]float64)
+	for _, typeID := range typeIDs {
+		priceHistory, _, err := b.esi.ESI.MarketApi.GetMarketsRegionIdHistory(ctx, regionID, typeID, nil)
+		if err != nil {
+			return nil, err
+		}
+		out[typeID] = movingAverage(7, priceHistory)
+	}
+	return out, nil
+}
+
+func movingAverage(days int, marketHistory []esi.GetMarketsRegionIdHistory200Ok) float64 {
+	historyLen := len(marketHistory)
+	if days > historyLen {
+		days = historyLen
+	}
+
+	var accum float64
+	for i := 1; i <= days; i++ {
+		dayData := marketHistory[historyLen-i]
+		accum += dayData.Average
+	}
+
+	return accum / float64(days)
+}
+
+func formatFuelPrices(blocks float64, fuelPrices map[int32]float64) string {
+	if fuelPrices == nil {
+		return "Error fetching prices."
+	}
+
+	// names must be indexed same as prices.
+	names := []string{
+		"[He]",
+		"[H]",
+		"[N]",
+		"[O]",
+	}
+	prices := []float64{
+		// Helium
+		fuelPrices[4247] * blocks,
+		// Hydrogen
+		fuelPrices[4246] * blocks,
+		// Nitrogen
+		fuelPrices[4051] * blocks,
+		// Oxygen
+		fuelPrices[4312] * blocks,
+	}
+	var minPriceIdx int
+	for i, price := range prices {
+		if price < prices[minPriceIdx] {
+			minPriceIdx = i
+		}
+	}
+
+	var b strings.Builder
+	b.WriteRune('\n')
+
+	for i, price := range prices {
+		// Lowest price is bold.
+		if i == minPriceIdx {
+			b.WriteString("**")
+		}
+		b.WriteString(names[i])
+		b.WriteRune(' ')
+		b.WriteString(humanize.CommafWithDigits(price, 0))
+		// Lowest price is bold.
+		if i == minPriceIdx {
+			b.WriteString("**")
+		}
+		b.WriteString(" ISK")
+		b.WriteRune('\n')
+	}
+
+	return b.String()
 }
 
 func structureByTypeID(typeID int32) structure {